Put reusable logic for other CSPs in abstract class
